+++ http4k-format-gson/src/main/kotlin/org/http4k/format/Gson.kt
@@ -2,46 +2,60 @@ package org.http4k.format
 
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonArray
+import com.google.gson.JsonDeserializationContext
+import com.google.gson.JsonDeserializer
 import com.google.gson.JsonElement
 import com.google.gson.JsonNull
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.google.gson.JsonPrimitive
+import com.google.gson.JsonSerializationContext
+import com.google.gson.JsonSerializer
 import org.http4k.core.Body
 import org.http4k.lens.BiDiBodyLensSpec
 import org.http4k.lens.BiDiWsMessageLensSpec
 import org.http4k.lens.ContentNegotiation
 import org.http4k.websocket.WsMessage
+import java.lang.reflect.Type
 import java.math.BigDecimal
 import java.math.BigInteger
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+import java.time.ZonedDateTime
+import java.time.format.DateTimeFormatter
+import java.util.UUID
 import kotlin.reflect.KClass
 
 class InvalidJsonException(messasge: String, cause: Throwable? = null) : Exception(messasge, cause)
 
 open class ConfigurableGson(builder: GsonBuilder) : JsonLibAutoMarshallingJson<JsonElement>() {
     override fun typeOf(value: JsonElement): JsonType =
-        when {
-            value.isJsonArray -> JsonType.Array
-            value.isJsonNull -> JsonType.Null
-            value.isJsonObject -> JsonType.Object
-            value.isJsonPrimitive -> {
-                val prim = value.asJsonPrimitive
-                when {
-                    prim.isBoolean -> JsonType.Boolean
-                    prim.isNumber -> JsonType.Number
-                    prim.isString -> JsonType.String
-                    else -> throw IllegalArgumentException("Don't know now to translate $value")
+            when {
+                value.isJsonArray -> JsonType.Array
+                value.isJsonNull -> JsonType.Null
+                value.isJsonObject -> JsonType.Object
+                value.isJsonPrimitive -> {
+                    val prim = value.asJsonPrimitive
+                    when {
+                        prim.isBoolean -> JsonType.Boolean
+                        prim.isNumber -> JsonType.Number
+                        prim.isString -> JsonType.String
+                        else -> throw IllegalArgumentException("Don't know now to translate $value")
+                    }
                 }
+                else -> throw IllegalArgumentException("Don't know now to translate $value")
             }
-            else -> throw IllegalArgumentException("Don't know now to translate $value")
-        }
 
     private val compact = builder.create()
     private val pretty = builder.setPrettyPrinting().create()
 
-    override fun String.asJsonObject(): JsonElement = JsonParser().parse(this).let { if(it.isJsonArray || it.isJsonObject) it else throw InvalidJsonException(
-        "Could not convert to a JSON Object or Array. $this"
-    ) }
+    override fun String.asJsonObject(): JsonElement = JsonParser().parse(this).let {
+        if (it.isJsonArray || it.isJsonObject) it else throw InvalidJsonException(
+                "Could not convert to a JSON Object or Array. $this"
+        )
+    }
+
     override fun String?.asJsonValue(): JsonElement = this?.let { JsonPrimitive(this) } ?: JsonNull.INSTANCE
     override fun Int?.asJsonValue(): JsonElement = this?.let { JsonPrimitive(this) } ?: JsonNull.INSTANCE
     override fun Double?.asJsonValue(): JsonElement = this?.let { JsonPrimitive(this) } ?: JsonNull.INSTANCE
@@ -82,4 +96,19 @@ open class ConfigurableGson(builder: GsonBuilder) : JsonLibAutoMarshallingJson<J
     inline fun <reified T : Any> WsMessage.Companion.auto(): BiDiWsMessageLensSpec<T> = WsMessage.json().map({ it.asA<T>() }, { it.asJsonObject() })
 }
 
-object Gson : ConfigurableGson(GsonBuilder().serializeNulls())
+object Gson : ConfigurableGson(GsonBuilder()
+        .registerTypeAdapter(LocalTime::class.java, custom({ LocalTime.parse(it, DateTimeFormatter.ISO_LOCAL_TIME) }, DateTimeFormatter.ISO_LOCAL_TIME::format))
+        .registerTypeAdapter(LocalDate::class.java, custom({ LocalDate.parse(it, DateTimeFormatter.ISO_DATE) }, DateTimeFormatter.ISO_DATE::format))
+        .registerTypeAdapter(LocalDateTime::class.java, custom({ LocalDateTime.parse(it, DateTimeFormatter.ISO_LOCAL_DATE_TIME) }, DateTimeFormatter.ISO_LOCAL_DATE_TIME::format))
+        .registerTypeAdapter(ZonedDateTime::class.java, custom({ ZonedDateTime.parse(it, DateTimeFormatter.ISO_ZONED_DATE_TIME) }, DateTimeFormatter.ISO_ZONED_DATE_TIME::format))
+        .registerTypeAdapter(UUID::class.java, custom(UUID::fromString))
+        .serializeNulls())
+
+private interface BidiJson<T> : JsonSerializer<T>, JsonDeserializer<T>
+
+private inline fun <T> custom(crossinline readFn: (String) -> T, crossinline writeFn: (T) -> String = { it.toString() }) =
+        object : BidiJson<T> {
+            override fun serialize(src: T, typeOfSrc: Type, context: JsonSerializationContext): JsonElement = JsonPrimitive(writeFn(src))
+
+            override fun deserialize(json: JsonElement, typeOfT: Type, context: JsonDeserializationContext): T = readFn(json.asString)
+        }