+++ http4k-format-moshi/src/main/kotlin/org/http4k/format/Moshi.kt
@@ -1,8 +1,10 @@
 package org.http4k.format
 
+import com.squareup.moshi.FromJson
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.KotlinJsonAdapterFactory
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.ToJson
 import org.http4k.core.Body
 import org.http4k.core.ContentType.Companion.APPLICATION_JSON
 import org.http4k.lens.BiDiBodyLensSpec
@@ -11,7 +13,14 @@ import org.http4k.lens.ContentNegotiation
 import org.http4k.lens.ContentNegotiation.Companion.None
 import org.http4k.lens.string
 import org.http4k.websocket.WsMessage
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+import java.time.ZonedDateTime
+import java.time.format.DateTimeFormatter
+import java.util.UUID
 import kotlin.reflect.KClass
+
 
 open class ConfigurableMoshi(builder: Moshi.Builder) : AutoMarshallingJson() {
 
@@ -28,9 +37,24 @@ open class ConfigurableMoshi(builder: Moshi.Builder) : AutoMarshallingJson() {
     inline fun <reified T : Any> asA(s: String): T = asA(s, T::class)
 
     inline fun <reified T : Any> Body.Companion.auto(description: String? = null, contentNegotiation: ContentNegotiation = None): BiDiBodyLensSpec<T> =
-        Body.string(APPLICATION_JSON, description, contentNegotiation).map({ asA(it, T::class) }, { asJsonString(it) })
+            Body.string(APPLICATION_JSON, description, contentNegotiation).map({ asA(it, T::class) }, { asJsonString(it) })
 
     inline fun <reified T : Any> WsMessage.Companion.auto(): BiDiWsMessageLensSpec<T> = WsMessage.string().map({ it.asA(T::class) }, { asJsonString(it) })
 }
 
-object Moshi : ConfigurableMoshi(Moshi.Builder().add(KotlinJsonAdapterFactory()))
+object Moshi : ConfigurableMoshi(Moshi.Builder()
+        .add(custom({ LocalTime.parse(it, DateTimeFormatter.ISO_LOCAL_TIME) }, DateTimeFormatter.ISO_LOCAL_TIME::format))
+        .add(custom({ LocalDate.parse(it, DateTimeFormatter.ISO_DATE) }, DateTimeFormatter.ISO_DATE::format))
+        .add(custom({ LocalDateTime.parse(it, DateTimeFormatter.ISO_LOCAL_DATE_TIME) }, DateTimeFormatter.ISO_LOCAL_DATE_TIME::format))
+        .add(custom({ ZonedDateTime.parse(it, DateTimeFormatter.ISO_ZONED_DATE_TIME) }, DateTimeFormatter.ISO_ZONED_DATE_TIME::format))
+        .add(custom(UUID::fromString))
+        .add(KotlinJsonAdapterFactory()))
+
+private inline fun <T> custom(crossinline readFn: (String) -> T, crossinline writeFn: (T) -> String = { it.toString() }) =
+        object {
+            @FromJson
+            fun read(t: String): T = readFn(t)
+
+            @ToJson
+            fun write(t: T): String = writeFn(t)
+        }